pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
 name = "wallpicker"
 version = "2.0.0"
 description = "Modern GTK4/Libadwaita wallpaper picker with Wallhaven integration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "pgotar", email = "pgotar@users.noreply.github.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Desktop Environment",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "PyGObject>=3.46.0",
    "send2trash>=1.8.2",
    "aiohttp>=3.9.0",
    "rapidfuzz>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "black>=24.0.0",
    "structlog>=24.0.0",
]

[project.scripts]
wallpicker = "main:main"

[project.urls]
Homepage = "https://github.com/gotar/wallpicker"
Repository = "https://github.com/gotar/wallpicker"
Issues = "https://github.com/gotar/wallpicker/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
wallpicker = ["data/*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["wallpicker"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["GObject.*", "Gio.*", "Gtk.*", "Gdk.*", "Adw.*", "GLib.*", "GdkPixbuf.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=.",
    "--cov-report=term-missing",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
asyncio_mode = "auto"
